feat: yuxrt Phase 2 — musl libm 移植 + mem 增强
- 从 musl-1.2.6 移植 28 个双精度数学函数（sin/cos/tan/exp/log/pow/sqrt 等）
- 6 个内部辅助函数（__cos/__sin/__tan/__rem_pio2/__rem_pio2_large/__expo2）
- 5 个错误处理函数（__math_divzero/invalid/oflow/uflow/xflow）
- 5 组数据表（exp_data/log_data/log2_data/pow_data/sqrt_data）
- 适配版 libm.h（无 CRT 依赖，yuxrt_/__yuxrt_ 命名约定）
- mem ops 从手写版替换为 musl 优化实现（memcpy/memmove/memset/memcmp）
- 产出 yuxrt.lib (322KB)，48 个新文件

+++ yuxrt/mem/mem.h
@@ -21,7 +21,7 @@ void yuxrt_free(void* ptr);
 void* yuxrt_memcpy(void* dst, const void* src, size_t size);
 
 // 将 dst 的前 size 字节设为 val。
-void* yuxrt_memset(void* dst, uint8_t val, size_t size);
+void* yuxrt_memset(void* dst, int val, size_t size);
 
 // 比较 a 和 b 的前 size 字节。
 // 返回 0 表示相等，<0 表示 a<b，>0 表示 a>b。
